(IMAGES-966) Automate Solaris 11.2 build process

+++ scripts/cleanup-aio.sh
@@ -5,23 +5,33 @@
 # as part of that process.
 
 # Uninstall the AIO
-if [[ ${PUPPET_AIO} == *".rpm"* ]] ; then
-  rpm -e puppet-agent
-elif [[ ${PUPPET_AIO} == *".deb"* ]] ; then
-  dpkg -P puppet-agent
-  #uninstall AIO for Solaris
-elif type pkg >/dev/null ; then
+if [ -n "${PC_REPO}" ] && [ "${PC_REPO}" != " " ]; then
+  if type pkg >/dev/null ; then
+  # Uninstall for Solaris using PC_REPO  
   pkg uninstall /system/management/puppet*
-  #uninstall hiera
+  # Uninstall hiera
   pkg uninstall pkg://solaris/library/ruby/hiera
-  # Uninstall puppet from macos
-elif [[ ${PUPPET_AIO} == *".dmg"* ]]; then
-  rm -rf /var/log/puppetlabs
-  rm -rf /var/run/puppetlabs 
-  pkgutil --forget com.puppetlabs.puppet-agent
-else
-  echo "Unsupported AIO package format" >&2
-  exit 1
+  else
+    echo "Unsupported AIO package format" >&2
+    exit 1
+  fi
+elif [ -n "${PUPPET_AIO}" ]; then
+  if [[ ${PUPPET_AIO} == *".rpm"* ]] ; then
+    rpm -e puppet-agent
+  elif [[ ${PUPPET_AIO} == *".deb"* ]] ; then
+    dpkg -P puppet-agent  
+  elif [ ${PUPPET_AIO} == *".dmg"* ] && [ "${PUPPET_AIO}" != " "]; then
+    # Uninstall puppet from macos
+    rm -rf /var/log/puppetlabs
+    rm -rf /var/run/puppetlabs 
+    pkgutil --forget com.puppetlabs.puppet-agent
+  elif [[ ${PUPPET_AIO} == *".p5p" ]]; then
+    # Uninstall puppet from Solaris using PUPPET_AIO    
+    pkg uninstall puppet-agent
+  else
+    echo "Unsupported AIO package format" >&2
+    exit 1
+  fi
 fi
 # Remove any Puppet-related files and directories
 rm -rf /etc/puppetlabs
